requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.1
django-filter==24.3
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.15.4
inflection==0.5.1
platformdirs==4.2.2
PyJWT==2.9.0
PyYAML==6.0.2
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
virtualenv==20.26.3